requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
six==1.15.0
toml==0.10.1
zipp==3.1.0
pillow